more hacker rank challenges

+++ src/main/java/RoadsAndLibariries.java
@@ -99,16 +99,16 @@ public class RoadsAndLibariries {
 		// calculate costs of repairing roads
 		for (Map.Entry<Integer, TreeSet<Integer>> entry : groupOfCities.entrySet()) {
 			// calculate cost of fixing roads
-			Integer size = entry.getValue().size();
-			Integer costOfFixingRoadsInThisEntry;
+			Long size = Long.valueOf(entry.getValue().size());
+			Long costOfFixingRoadsInThisEntry;
 			if (size > 1) {
 				costOfFixingRoadsInThisEntry = ((size - 1) * c_road) + c_lib;
 			} else {
 				// if city is isolated, then there is no way but building a new lib
-				costOfFixingRoadsInThisEntry = Integer.MAX_VALUE;
+				costOfFixingRoadsInThisEntry = Long.MAX_VALUE;
 			}
 			// calculate cost of building lib in each city
-			Integer costOfLibInEachCity = size * c_lib;
+			Long costOfLibInEachCity = size * c_lib;
 			totalMinumum = totalMinumum + Math.min(costOfFixingRoadsInThisEntry, costOfLibInEachCity);
 			totalCitiesAppeared = totalCitiesAppeared + size;
 		}
